Customize some Rust-mode aspects for Emacs.
Uses of rustfmt no longer pass --edition because that conflicts with a project's
.rustfmt.toml having a different value for the edition (e.g. projects that need
to support old versions).  Now, just rely on projects' choices, instead.

Uses of cargo now pass --all-targets (the default, like before), because I now
have a trick for avoiding build errors with benches/ targets with stable rustc
versions, and because trying to pass a fixed selection of targets doesn't work
across various projects with different kinds of targets.

There are also a couple changes to LSP options which sound like improvements and
seem to not be causing issues, but I'm not certain yet.