Don't define programs.firefox.extensions and remove my.rycee option.
The URL to download tree-style-tab was broken today from the firefox-addons of
rycee.  This caused `home-manager build` to fail.  I don't want that to happen
anymore.

Defining programs.firefox.extensions was ineffective anyway, because Firefox
still manages the extensions' updates even with that.  So, it's not needed and
was just causing the above problem.