Use latest Magit of MELPA, instead of last release of MELPA Stable.
Primarily because the last stable release is old enough that it doesn't support
recent changes to Git now.  Also because I've noticed a couple new features that
are nice to have in the latest Magit.

I'd used the stable Magit before because I wanted rock-solid stability and was
concerned that the latest could cause instability that is unpredictable due to
continuously rolling updates.  But after trying the latest for a couple months
and going through a few updates with it, it seems stable enough, and it has far
more downloads/users so it seems to be the de-facto official version.