For TRAMP, don't use absolute path to Bash executable.
Some OSs, e.g. FreeBSD, do not have it at /bin/bash but have it somewhere else.
Just invoke "bash" to allow it to be found wherever in the user's PATH.

It's also necessary to adjust tramp-sh-extra-args, because the default value for
that was designed poorly - it should not assume that the shell executable names
will always be preceded by a `/` character.  Unfortunately, now I'll have to
maintain my customization of this variable too now, just to fix it when all I
want is for the default to work.