Start following the XDG Base Directory Specification.
Not all of my setup does so yet.

Move my shell customizations to be XDG-BDS compliant.

This has the benefits of:

- Using the different XDG-BDS locations, which can have functional/operational
  benefits (e.g. my lock file for my Bash combined history is now in the
  "runtime" location, which would be needed when a home directory is on a shared
  file-system).

- Respecting the XDG-BDS user-specific environment variables if they're set.

- Uncluttering top-level home directories.

- Better organization of the files.

This also introduces some more MY_BASH_* variables in Bash sessions, and renames
previous MY_* to MY_BASH_*.  Most of these are not exported to the environment.

This commit introduces the "my/" sub-dir namespace under the XDG-BDS locations,
for a user's private use.  This should never conflict with any other
application, because, even though there are some apps whose name begins with
"my", there are not (to my knowledge) any that are named only that and there
should never be (because that would be a ridiculous name for a published app).

Revise comments to mention $XDG_CONFIG_HOME/ instead of ~/.config/.

Extend my-ibuffer-project-root-regexps for XDG-BDS dirs, for Emacs.