Add support for deploying and using my `dotfiles` repo in other OSs.
This is desirable for more-easily having my personal setup in other hosts I work with.  E.g. VMs
for development or company datacenters with home directories shared across multiple machines.

Previously, it assumed it'd only be used with Home Manager in NixOS as configured by my
`nixos-config` repository and with only a single CPU architecture.

Now, it can work in any POSIX OS, with only minimal common portable utilities needed to deploy it
to such, and an instance of it can work when simultaneously shared across multiple different OSs
with different CPU architectures.  It tries to configure things to be the same or similar in
different OSs.

There is a new utility that conveniently automates the deployment of it to any remote user's home
directory.  This merges the remote user's preexisting files with those from my `dotfiles`, and
this preserves the relation to both in the repository history.

To have a consistent set of tools be installed in multiple OSs, it declares which packages to
install via a generic single list, and it has platform-specific procedures for actualizing this in
each.

It has a pathname scheme for uniformly and hierarchically locating platform-specific and
architecture-specific parts, based on a scheme for identifying different platforms.

It tries to gracefully fail when some parts of it can't be setup (e.g. due to lack of
permissions), so that at least some of it can be used immediately as a remote user.

It is designed to have support for further OSs added incrementally in the future, as needed.

There is a Vagrantfile for creating VMs of various OSs with automatic deployment of it to each.

Some initialization files now have some conditional logic to adjust for when running in a specific
platform.  Further such adjusting probably still needs to be added to some other init & config
files.

The GNU Core Utilities are configured to be used by default in my shell sessions, in preference to
other utilities, for non-GNU OSs like FreeBSD & Solaris.  This is only because I'm more familiar
with GNU's and want consistency with my predominant use of Linux.  These are used via shell
functions that are limited in scope, and not via changing the PATH environment variable (except
for Emacs' TRAMP), so that a host's native utilities are still used by its things.

There is now a small collection of helper functions for POSIX-Shell and Bash.  These are used by
the new facility and are also now provided in interactive sessions.

The changes to the .nix files are mostly to allow them to work in other NixOS hosts that are
configured differently than my /etc/nixos/configuration.nix.

A few of my simple personal utilities are now moved to be preinstalled for use in any
OS.  (Before, they were only installable via the Nix package manager.)

This commit also incorporates some miscellaneous tweaks I'd made for myself in the past when
working with different OSs.

This commit also has a little clean-up of a few prior parts.

This commit constitutes over 100 squash and fixup commits.