Fix use of tput for my Bash prompt, for FreeBSD and others.
Unlike OpenIndiana, FreeBSD does not provide a "GNU" version of `tput`.  FreeBSD's default `tput`
is that of ncurses, like GNU OSs, but Solaris has its own as default and so its alternative under
/usr/gnu/ must instead be used which is its ncurses one.

The goal is to consistently use the ncurses `tput` because it uses ncurses' better support for a
larger variety of $TERM types.