Minor fixes for my Linux/Alpine packages.
Have the `man-db` package spec.
I'd forgotten to add it along with the `man-pages` package.

Install the `tput` utility.
This was not needed before and is only recently needed for the new improvements to my Bash prompt.

Install Clippy for Rust+Cargo.
My rust-analyzer config uses it and assumes it's available.
Alpine doesn't include it in its `cargo` package, unlike most other distros.