Improve my-deploy-setup.
Default to using the main branch for merging into the target user's, because I've found that this
is more desirable more often.  Previously, it defaulted to using the invoking user's branch, but
oftentimes it's undesirable to expose that branch in the target user's security domain, and using
that branch often causes merge conflicts which is also undesirable.  Now, the invoking user must
explicitly supply their branch, via the MY_DEPLOY_SETUP_DOTFILES_FROM_REFSPECS environment
variable, or via pushing that branch later, if desired.

Handle recovery from merge failures better, by having the option to reattempt the merge again to
resolve conflicts and continue the repository setup like normal, which is easier for the user
encountering such failure to understand how to deal with (versus needing to know the raw commands
that should be done).  The previous raw recovery approach is still an option.

Change the left-over `preexisting` and `deployed-from` refs to be tags instead of branches,
because they are supposed to be only unchanging markers.  Like previously, they start-out as
branches as appropriate for the internal operations of the script, and the raw recovery approach,
if needed, can still change the `deployed-from` branch before it's finalized.