Adjust Emacs init for versions 27 and 28, which are less-than 29.
Add many uses of `:ensure nil` to the `use-package` forms that are for built-in libraries, because
this is needed for Emacs 27 because its `package-installed-p` returns `nil` for built-in
libraries (in contrast to 29's which returns `t`).  I don't know if this is also the case with 28.
Adding these feels uglier, but I'd already previously needed to add this to some of the other
forms to avoid some other issue even with 29.

The loading of the `my-platform` library is now done with `require`, because `use-package` is not
built-in to Emacs versions less-than 29, and, even when it's been installed for them, things are
not yet setup to find ELPA packages during the `early-init.el`.

The `require`ing of `use-package` is now allowed to fail, for when it's not yet installed for
Emacs versions less-than 29.  When it fails, all uses of `use-package' are now ignored, because
such reduced functionality of having whatever else was still setup is better than having errors
when executing the init file.

The byte-compiling of the "my" directory is now done only when `require`ing `use-package`
succeeds, because many of the `.el`s in that directory reference `use-package` and byte-compiling
them when it's not loaded (because it's not yet installed) results in flawed `.elc`s.  This
byte-compiling is now done with demoted errors, because any errors from that should not prevent
executing the rest of the init file, because it's not necessary to have the `.elc`s.

The setting of options in `xdg-bds.el` is now done with `eval-after-load`, so this still works
when `use-package' is unavailable.

The expressions that depend on `use-package` forms working should now all be inside those
respective forms, so that they are ignored along with the forms when `use-package` is unavailable.

The other changes are only minor, to use the equivalent functions of older Emacs versions.