For FreeBSD and OpenIndiana, have GNU utils in PATH for Emacs.
Because Emacs runs commands, e.g. `ls` and `diff`, assuming they are the GNU ones.  Both Emacs
libraries and my own Emacs settings do such assuming, and this breaks functionality when the
non-GNU default/stock utils of a platform are incompatible with GNU-style arguments/behavior.
This would still happen even without my own settings, because such assumptions are hardcoded in
parts that are built-in to or in common ELPA libraries for Emacs.

This should also be done for any other non-GNU OSs for which support is added to this repo in the
future.

Hopefully, this will not cause new problems with a platform's other programs, if those are run by
Emacs such that the emacs process's environment is inherited.  If those need the utils that are
named the same to be the platform's non-GNU default/stock ones because the platform's have
non-POSIX behavior that the GNU's don't, then there could hypothetically be a problem.  But
risking that unlikelihood seems better than certainly having broken frequently-desired Emacs
functionality.

A user can very easily disable this use of GNU utils by simply not using this `emacs` symlink in
their PATH, either by temporarily arranging their PATH such, or by permanently removing the
symlink which is located in ~/bin/my/platform/$MY_PLATFORM_OS/**/bin/.