Adjust Emacs init for version 28 with native compilation.
Fix my eln-cache dir replacing to not use startup-redirect-eln-cache.  I didn't catch this in my
previous testing because I'd tested in Ubuntu 22.04, Alpine 3.18, and Ubuntu 23.10 which have
Emacs 27 w/o native-comp, 28 w/o native-comp, and 29 w/ native-comp, respectively, all of which
avoided the bug (because either native-comp is absent or the startup-redirect-eln-cache function
exists), but I hadn't tested in Debian 12, until now, which has Emacs 28 w/ native-comp w/o that
function.

Fix my/lib/*.el to require `use-package` as needed for native compilation of them by Emacs 28
which doesn't have `use-package` built-in.  Native compilation is done in subprocesses in batch
mode and as such the requiring of `use-package` in my init file is not done, and so it must be
done in each file in order for the compilation to know that the `use-package` forms are
macros (not function calls).  Otherwise, flawed compilations were created that caused errors when
their code ran.