Improve my Emacs init w.r.t. compilation and batch mode.
Prevent native compilation during early-init.el.  For some platforms, e.g. Debian 12, it was
native-compiling pcase.el into the undesirable ~/.config/emacs/eln-cache/ dir because it loads
pcase before it replaces that dir and it doesn't already have a .eln.  This should also prevent
such possibility for other libraries in the future.

Activate all packages when in `--batch --user` mode (which is used by my-install-emacs-packages),
because batch mode doesn't do this (unlike normal mode which does) and this could be needed by
uses that assume things are initialized like normal mode because the user init file was loaded.

Improve the re-prepending of my-eln-cache-dir so that it's done more cleanly - only when needed
and remove any other occurrence.

Don't compile my/lib/my-platform.el, to prevent the possibility of its .elc becoming outdated but
still being loaded instead of its .el when the .el was changed, because it's loaded by
early-init.el before init.el recompiles my/lib/ (before the rest are loaded).  It's the only
my/**/*.el that this is a concern for (because it's the only one loaded by early-init.el).