Encourage users to track Nix state files, by not ignoring them.
NixOS 23.11 changes the nix-channel utility to locate its channel generations in
$XDG_STATE_HOME/nix/profiles/ and this is enabled by default.  23.11 also introduces further new
changes to locate other per-user Nix files according to XDG-BDS but these are still experimental
and disabled by default.  Because these critical files will be changed around in the future, they
should be tracked but in each user's per-user private branch because the contents are per-user.

I'd previously ignored them when I was new to NixOS and wasn't sure, but now I know their contents
rarely change and that it will be good to track them.