Ignore future ~/.local/state/nix/profiles/.
NixOS 23.11 changes its utilities to locate their "profile generations" here.  These should not be
tracked because their file timestamps are important (especially for GC'ing them) and changing the
working-tree checkout of this repository (which is done occasionally, e.g. switching to the main
branch temporarily to do some operation on it) would invalidate those timestamps (when those
files (symlinks, actually) are recreated by the checkout changing).

Further, new symlinks are continously created, and old ones deleted, as the profiles are updated
by Nix operations, which would be tedious to be having to continously make commits for.

Finally, there wouldn't be much value in tracking the history of such changes, because these
profile generations are meaningful in the host that they're for and during the time period that
they're still alive, but this repository is intended for tracking a user's changes that should be
meaningful when the user's home is moved to or recreated in a different host.

The parent directory, ~/.local/state/nix/, is still not ignored, because other further-future
NixOS versions seem like they're planning on moving the ~/.nix-* files under there and I'll still
want to track them there.