Add config for KeePassXC, including as Secret Service.
There is a corresponding commit in my nixos-config repo that:
installs the KeePassXC package system-wide, and
allows Firefox to do native-messaging with it, and
assumes it's config'ed to provide the Secret Service API (per this commit), and
enables the git-credential-libsecret helper (which will use KeePassXC), and
enables the ssh-agent (which KeePassXC can integrate with).

The browser extension must be manually installed and configured, because formalizing and
automating that is too difficult for now.  But maybe that could be done, at least to some extent,
in the future, by utilizing the newer options for Firefox (and (Ungoogled) Chromium?) that Home
Manager and NixOS have.

Add a file of exported KeePassXC-Browser settings, for users to start with.  This file is
inactive, not used, and to be effective each user must import it into the KeePassXC-Browser
extension (after installing that in a web browser, of course).

The old config to use the git-credential-cache helper is deleted, so that the now-config'ed
git-credential-libsecret helper will be used instead.

The KeePass-DB files are ignored w.r.t. tracking in this repo, because they're
frequently-changing, large binary encrypted gibberish, too sensitive and private (even as
encrypted), and should be preserved via separate backups.

Add a KeePass-DB file as a template (not ignored, as a special exception), for users to start
with.  This file is inactive, not used, and is only effective if a user chooses to copy/merge it
to make an active DB file (which will be ignored).  The location and name of this template file is
only to organize it mnemonically for when users want to find it to start a separate DB.  The
password for this template DB is "password", because users need to be able to unlock it to use it
to start their separate DBs.  As such, it is critical that each user changes the password for
their separate DB, e.g. to a high-entropy passphrase.

For non-NixOS OSs, all this config for KeePassXC is inactive.  This is desired, because my setup
for other OSs doesn't assume GUI desktop nor Linux, and so shouldn't assume that KeePassXC could
be used (which requires GUI) nor as Secret Service (which is Linux-specific (but could
hypothetically, and perhaps is already, implemented for some other OS, but can't assume that it
will be for all POSIX OSs)).  However, if KeePassXC were installed (maybe just its keepassxc-cli)
in some other OS, then most of this config and files would be useful.

For non-NixOS OSs, because my global Git config no longer has git-credential-cache helper, and
without git-credential-libsecret helper in an OS which doesn't have that, usage of Git will no
longer use any credential helper (unless some helper is manually configured there).  This is
probably for the best anyway, because my setup for other OSs shouldn't assume how this should be
in various other situations.