For pushing to GitHub, do so via SSH, when pulling is via HTTPS.
This is more convenient, because it enables remotes to have HTTPS as their primary URL (which
doesn't require authentication to do fetching/pulling) but to use SSH for pushing (which doesn't
require asking for credentials (when the SSH key is already added to ssh-agent, which my KeePassXC
configs usually do automatically)), and SSH keys don't expire (unlike GitHub's access tokens).

The public keys of GitHub are now supplied by ~/.ssh/known_hosts for all users, because this fits
with now defaulting to using SSH with that site.  These should only very rarely change, and so
won't be a bother in the dotfiles repo, and when they do change it'll be desirable to supply that
to all users by a single commit for them to merge.  GitHub is so widely used for so many things
nowadays, and so providing these for users might end-up being helpful for other purposes.  I
verified that these are what GitHub's official documentation (loaded via TLS) lists.

Note that users' private branches can still disable or otherwise change any of this, if desired.
The additions of this commit could be deleted.  Or Git's remote.<name>.pushurl config option could
be used to override.  Or remote.<name>.url could be a non-matching URL, to avoid matching this
commit's pushInsteadOf value.

For other Git-hosting platforms like GitLab, a similar setup like this commit could probably also
be added in the future, if desired.