Improve file-mode handling of my Bash interactive history files.
If a previous combined history file (which is supposed to only be temporary briefly) is left
around due to a crash, try to make it be read-only.

If the primary combined history file somehow becomes read-only, try to make it writable again.

These changes are a little more robust and so might as well be done.

I've had some previous combined history files left around, and I still don't understand what
caused that, and I tried to reproduce, and the only thing that seems like it can cause that is for
a bash process to immediately be killed (i.e. SIGKILL, SIGSEGV, etc).  At least now, such
left-around files will be read-only for when I later notice them and can maybe see if they offer
any clue as to when and why a bash process crashed.