Add custom prompt for GDB (when run in console).
It is ignored when run via MI mode (e.g. under Emacs).

It required removing my automatic enabling of TUI mode, because there was some bug with having a
custom prompt and with the Curses displaying when TUI mode starts immediately with my layout.  But
this bug doesn't occur if TUI mode is started without my layout and then my layout is chosen
after, so I'll just go with that so I can have my custom prompt and custom layout both working.