Fix my Bash interactive history combining when terminal is closed.
Otherwise, previously, when a terminal is closed directly, the SIGHUP that sends to the associated
`bash` process would cause the session's history to not be recorded in the combined history file,
because of the interplay between how Bash handles that signal and its EXIT trap.

This fix is somewhat of a hack, but I think it'll remain reliable into the farther future, based
on my debugger stepping of this issue showing what Bash does differently with this fix.