Improve Emacs startup time for non-NixOS OSs.
By avoiding doing `package-refresh-contents`.  Previously, it was being done due to some
`use-package` ensuring for libraries that are both built-in and provided by the package archives.
It's unnecessary to ensure those are installed as packages, because they're already built-in.  I
wasn't being annoyed by this often, because in my NixOS hosts my `package-archives` is `nil` which
caused the refreshing to finish immediately, but in other non-NixOS hosts my `package-archives` is
non-`nil` and when the network requests are not super fast it was slowing down Emacs' startup
significantly.