Improve ClangD config.
The `-W` flags are changed to: add many that I think will be desirable (based on my exhaustive
review of all of the diagnostic flags of Clang 19.1); and remove those that are GCC-only and so
were being silently ignored.

The `lsp-clangd-version` of Emacs is changed to a bogus version number, only because I don't want
the automatic installing of clangd (which I never was doing anyway).

Two additional command-line arguments to `clangd` are now used, for Emacs, that I think will be
desirable.

The `lsp-clients-clangd-executable` of Emacs is changed to use my wrapper that passes
`--query-driver`.  This is cleaner than just having that wrapper be the `clangd` in PATH, because
some other non-Emacs things that use PATH might not always want the wrapper's effects.

The `--query-driver` value passed by my clangd wrapper now includes compilers under
~/.nix-profile/bin/ and /usr/local/bin/, because sometimes one of those pathname prefixes is
what's needed.  Those locations are trustworthy.