Install a file-and-dir monitoring package, for the various OSs.
My current need for this is for using TRAMP to access my remote home dirs (otherwise, TRAMP
endlessly complains (if you don't disable the Emacs setting) about not having a file-notification
program, and having it is indeed better).

Having this might also be useful for other things in the future.