Adjust custom GDB windows layout for Emacs GUD.
This allows it to work when Emacs is running in non-GUI terminal.  Otherwise, the pixel values
were causing `window--state-put-1` to error and so mess-up the layout.  The pixel values seem to
be unnecessary to have the same layout and proportions as given by the line, column, and fraction
values.  Some of the `normal-height` fraction values seem to be unnecessary and are removed
because they'd need to be different for these changes but I didn't calculate what they'd need to
be.

Also, increase the lower windows' size, and decrease the uppers', by one line, because I prefer
that and that's better when in terminal where it's more difficult to resize windows.